+++ changelog.md
@@ -1,4 +1,11 @@
 # Changelog
 
-**./scripts/text-styling.js**
-* Added comments.
+**./script/definitions.js - getSymbolDefinitions**
+* Renamed 'symbol' properties to 'glyph'
+
+---
+
+**./script/render-common.js - chooseSymbols**
+* Changed 'listSymbols.symbol' reference to 'listSymbols.glyph'
+* Renamed 'unicodeSymbols' variable to 'glyphSymbols'
+* Changed "Unicode symbols." comment to "Glyph symbols."